requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
python-dotenv==0.19.0
reportlab==3.6.8
PyMySQL==1.0.2
gunicorn==20.1.0
mysqlclient==2.1.1
